Add database

+++ src/main/java/ru/yandex/practicum/filmorate/exception/FilmNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.exception;
 
 public class FilmNotFoundException extends RuntimeException {
-    public FilmNotFoundException(int id) {
+    public FilmNotFoundException(long id) {
         super(String.format("Film with id = %s is not found.", id));
     }
 }
