+++ src/main/java/ru/yandex/practicum/filmorate/exception/NoValidIdException.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.exception;
 
 public class NoValidIdException extends RuntimeException {
-    public NoValidIdException(int id) {
+    public NoValidIdException(long id) {
         super(String.format("Id = %s is incorrect.", id));
     }
 }
