+++ src/main/java/ru/yandex/practicum/filmorate/exception/UserLikeNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.exception;
 
 public class UserLikeNotFoundException extends RuntimeException {
-    public UserLikeNotFoundException(int filmId, int userId)  {
+    public UserLikeNotFoundException(long filmId, long userId)  {
         super(String.format("User with id = %s like on film with id = %s is not found.", userId, filmId));
     }
 }
