+++ src/main/java/ru/yandex/practicum/filmorate/exception/UserNotFoundException.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.exception;
 
 public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(int id) {
+    public UserNotFoundException(long id) {
         super(String.format("User with id = %s is not found.", id));
     }
 }
